+++ CHANGELOG.md
@@ -36,6 +36,9 @@
 * [UIQM-499](https://issues.folio.org/browse/UIQM-499) Auto-populate empty positions for fields 006, 007 and 008.
 * [UIQM-380](https://issues.folio.org/browse/UIQM-380) Create Orig Authority Record: Open new record in quickMARC UI.
 * [UIQM-326](https://issues.folio.org/browse/UIQM-326) Add auto-linking MARC bib fields.
+* [UIQM-443](https://issues.folio.org/browse/UIQM-443) Auto-linking MARC bib. Handling multiple authority records matches.
+* [UIQM-436](https://issues.folio.org/browse/UIQM-436) Auto-linking MARC bib. Handling multiple subfield 0.
+* [UIQM-509](https://issues.folio.org/browse/UIQM-509) Autopopulate subfields when Creating MARC records.
 
 ## [6.0.2](https://github.com/folio-org/ui-quick-marc/tree/v6.0.2) (2023-03-30)